+++ src/Core/Checker.cpp
@@ -127,6 +127,7 @@ void Checker::prepare(const QString &compileCommand)
         connect(compiler, SIGNAL(compilationFinished(const QString &)), this, SLOT(onCompilationFinished()));
         connect(compiler, SIGNAL(compilationErrorOccurred(const QString &)), this,
                 SLOT(onCompilationErrorOccurred(const QString &)));
+        connect(compiler, SIGNAL(compilationFailed(const QString &)), this, SLOT(onCompilationFailed(const QString &)));
         connect(compiler, SIGNAL(compilationKilled()), this, SLOT(onCompilationKilled()));
         compiler->start(checkerPath, "", compileCommand, "C++");
     }
@@ -167,6 +168,11 @@ void Checker::onCompilationFinished()
 void Checker::onCompilationErrorOccurred(const QString &error)
 {
     log->error(tr("Checker"), tr("Error occurred while compiling the checker:\n%1").arg(error));
+}
+
+void Checker::onCompilationFailed(const QString &reason)
+{
+    log->error(tr("Checker"), tr("Failed to compile the checker: %1").arg(reason), false);
 }
 
 void Checker::onCompilationKilled()
@@ -217,7 +223,7 @@ void Checker::onRunFinished(int index, const QString &, const QString &err, int 
 
 void Checker::onFailedToStartRun(int index, const QString &error)
 {
-    log->error(head(index), error);
+    log->error(head(index), error, false);
 }
 
 void Checker::onRunOutputLimitExceeded(int index, const QString &type)
@@ -229,7 +235,8 @@ void Checker::onRunOutputLimitExceeded(int index, const QString &type)
             .arg(type)
             .arg(index + 1)
             .arg(SettingsHelper::getOutputLengthLimit())
-            .arg(SettingsHelper::pathOfOutputLengthLimit()));
+            .arg(SettingsHelper::pathOfOutputLengthLimit()),
+        false);
 }
 
 void Checker::onRunKilled(int index)